requirements.txt
pytorch=2.1.0
torchtext=0.6.0
pathspec=0.12.1